+++ src/main/java/com/tsp/new_tsp_front/api/agency/domain/FrontAgencyEntity.java
@@ -1,6 +1,5 @@
 package com.tsp.new_tsp_front.api.agency.domain;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.tsp.new_tsp_front.api.common.domain.CommonImageEntity;
 import com.tsp.new_tsp_front.api.common.domain.NewCommonMappedClass;
 import com.tsp.new_tsp_front.api.model.domain.FrontModelEntity;
@@ -52,13 +51,11 @@ public class FrontAgencyEntity extends NewCommonMappedClass {
     @NotEmpty(message = "소속사 노출 여부 선택은 필수입니다.")
     private String visible;
 
-    @JsonIgnore
     @BatchSize(size = 5)
     @Where(clause = "type_name = 'agency'")
     @OneToMany(mappedBy = "frontAgencyEntity")
     private List<CommonImageEntity> commonImageEntityList = new ArrayList<>();
 
-    @JsonIgnore
     @OneToOne(mappedBy = "frontAgencyEntity", cascade = ALL, fetch = LAZY)
     private FrontModelEntity frontModelEntity;
