+++ src/main/java/com/tsp/new_tsp_front/api/model/domain/FrontModelEntity.java
@@ -1,6 +1,5 @@
 package com.tsp.new_tsp_front.api.model.domain;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.tsp.new_tsp_front.api.agency.domain.FrontAgencyEntity;
 import com.tsp.new_tsp_front.api.common.domain.CommonImageEntity;
 import com.tsp.new_tsp_front.api.common.domain.NewCodeEntity;
@@ -128,30 +127,25 @@ public class FrontModelEntity extends NewCommonMappedClass {
     @Convert(converter = CustomConverter.class)
     private ArrayList<CareerJson> careerList;
 
-    @JsonIgnore
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "category_cd", insertable = false, updatable = false)
     private NewCodeEntity newModelCodeJpaDTO;
 
     @Builder.Default
-    @JsonIgnore
     @BatchSize(size = 5)
     @Where(clause = "type_name = 'model'")
     @OneToMany(mappedBy = "frontModelEntity", fetch = LAZY, cascade = REMOVE)
     private List<CommonImageEntity> commonImageEntityList = new ArrayList<>();
 
-    @JsonIgnore
     @OneToOne(fetch = LAZY, cascade = ALL)
     @JoinColumn(name = "agency_idx", referencedColumnName = "idx", insertable = false, updatable = false)
     private FrontAgencyEntity frontAgencyEntity;
 
     @Builder.Default
-    @JsonIgnore
     @OneToMany(mappedBy = "frontModelEntity", fetch = LAZY)
     private List<FrontScheduleEntity> modelScheduleList = new ArrayList<>();
 
     @Builder.Default
-    @JsonIgnore
     @OneToMany(mappedBy = "frontModelEntity", fetch = LAZY)
     private List<FrontNegotiationEntity> modelNegotiationList = new ArrayList<>();
 
@@ -160,6 +154,11 @@ public class FrontModelEntity extends NewCommonMappedClass {
     }
     public void updateFavoriteCount() {
         this.modelFavoriteCount++;
+    }
+
+    public void addNegotiation(FrontNegotiationEntity frontNegotiationEntity) {
+        frontNegotiationEntity.setFrontModelEntity(this);
+        this.modelNegotiationList.add(frontNegotiationEntity);
     }
 
     public static FrontModelDTO toDto(FrontModelEntity entity) {
